created a new class for dice random method
